// autogpt_platform/frontend/src/app/api/__generated__/models/storeAgentDetails.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { StoreAgentDetailsInstructions } from "./storeAgentDetailsInstructions";
import type { StoreAgentDetailsRecommendedScheduleCron } from "./storeAgentDetailsRecommendedScheduleCron";
import type { StoreAgentDetailsChangelog } from "./storeAgentDetailsChangelog";

export interface StoreAgentDetails {
  store_listing_version_id: string;
  slug: string;
  agent_name: string;
  agent_video: string;
  agent_output_demo: string;
  agent_image: string[];
  creator: string;
  creator_avatar: string;
  sub_heading: string;
  description: string;
  instructions?: StoreAgentDetailsInstructions;
  categories: string[];
  runs: number;
  rating: number;
  versions: string[];
  graph_id: string;
  graph_versions: string[];
  last_updated: Date;
  recommended_schedule_cron?: StoreAgentDetailsRecommendedScheduleCron;
  active_version_id: string;
  has_approved_version: boolean;
  changelog?: StoreAgentDetailsChangelog;
}